Add files via upload
Learning to connect the python environment with the brokerage account was quite confusing, but with the assistance of stackoverflow community and IBKR API, I've managed to successfully retrieve my first data from the terminal. First steps towards backtesting, and the next will be to understand the IB interface and to focus on active portfolio management for now.